Add replsetConf metrics
We already pull the replsetStatus metrics, but things such as priority,
hidden, and votes are in the conf. This exposes them. For now the tags
mimic what is in mongo (so they match the docs) -- sadly this does mean
that the tags don't match what the other replset metrics have (set ->
id, etc.)